Create  Longest-common-prefix
In this solution, I compared characters of all strings one by one starting from the first character. I used the first string as a reference and checked whether the same characters appear at the same position in the other strings. When a mismatch occurs, I stop and return the part of the string that matched so far as the longest common prefix.