Create Greatest no of  Candies
This program checks which kids can have the greatest number of candies if we give some extra candies.
First, it finds the kid with the most candies. Then for each kid, it adds the extra candies and sees if they can reach or exceed the maximum. The program returns true if yes, otherwise false.